Enhanced handlers with classes
